Move cGame on game/
